app/requirements.txt
fastapi==0.121.3
uvicorn[standard]==0.38.0
pymysql==1.1.2
python-dotenv==1.2.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pycryptodome==3.23.0
tencentcloud-sdk-python==3.0.1491